Refactor camerapose (#31)
* Removed alpha from CameraPose. TODO update pybinds

* Updated pybind to remove alpha from CameraPose

* Massive refactoring to change CameraPose to use quaternion instead of rotation matrix

* Fix bug with inconsistent compilation options for pybind

* Added inline to CameraPose helper functions

* Added in ffast-math but removed fun-safe-math-opt

* Update compilation options. Removed removed warnings and fixed errors.

* Restructuring

* replace pragma once with header guards

* Rename PROSAC option and changed default to false

* missing update

* update comments

* Added missing copyright notices

+++ PoseLib/robust/sampling.cc
@@ -1,3 +1,33 @@
+// Copyright (c) 2021, Viktor Larsson
+// All rights reserved.
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions are met:
+//
+//     * Redistributions of source code must retain the above copyright
+//       notice, this list of conditions and the following disclaimer.
+//
+//     * Redistributions in binary form must reproduce the above copyright
+//       notice, this list of conditions and the following disclaimer in the
+//       documentation and/or other materials provided with the distribution.
+//
+//     * Neither the name of the copyright holder nor the
+//       names of its contributors may be used to endorse or promote products
+//       derived from this software without specific prior written permission.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+// ARE DISCLAIMED. IN NO EVENT SHALL COPYRIGHT HOLDERS OR CONTRIBUTORS BE LIABLE
+// FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+// (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+// LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+// ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+// (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+// SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+
+#include <cmath>
 #include "sampling.h"
 
 namespace pose_lib {
@@ -14,13 +44,13 @@ int random_int(RNG_t &state) {
 
 // Draws a random sample
 void draw_sample(size_t sample_sz, size_t N, std::vector<size_t> *sample, RNG_t &rng) {
-    for (int i = 0; i < sample_sz; ++i) {
+    for (size_t i = 0; i < sample_sz; ++i) {
         bool done = false;
         while (!done) {
             (*sample)[i] = random_int(rng) % N;
 
             done = true;
-            for (int j = 0; j < i; ++j) {
+            for (size_t j = 0; j < i; ++j) {
                 if ((*sample)[i] == (*sample)[j]) {
                     done = false;
                     break;
@@ -31,7 +61,7 @@ void draw_sample(size_t sample_sz, size_t N, std::vector<size_t> *sample, RNG_t 
 }
 // Sampling for multi-camera systems
 void draw_sample(size_t sample_sz, const std::vector<size_t> &N, std::vector<std::pair<size_t, size_t>> *sample, RNG_t &rng) {
-    for (int i = 0; i < sample_sz; ++i) {
+    for (size_t i = 0; i < sample_sz; ++i) {
         bool done = false;
         while (!done) {
             (*sample)[i].first = random_int(rng) % N.size();
@@ -41,7 +71,7 @@ void draw_sample(size_t sample_sz, const std::vector<size_t> &N, std::vector<std
             (*sample)[i].second = random_int(rng) % N[(*sample)[i].first];
 
             done = true;
-            for (int j = 0; j < i; ++j) {
+            for (size_t j = 0; j < i; ++j) {
                 if ((*sample)[i] == (*sample)[j]) {
                     done = false;
                     break;
